+++ src/gadgets.rs
@@ -8,16 +8,23 @@
 //!
 //! This module provides Plonk gadgets for verification of Schnorr signatures.
 
-use dusk_jubjub::{GENERATOR_EXTENDED, GENERATOR_NUMS_EXTENDED};
+use dusk_jubjub::GENERATOR_EXTENDED;
 use dusk_poseidon::sponge;
+
+#[cfg(feature = "double")]
+use dusk_jubjub::GENERATOR_NUMS_EXTENDED;
 
 use dusk_plonk::prelude::*;
 
-/// Verifies a single-key Schnorr signature within a Plonk circuit without
-/// requiring the secret key as a witness.
+/// Verifies a single-key Schnorr signature [`Signature`]within a Plonk circuit
+/// without requiring the secret key as a witness.
 ///
 /// The function performs Schnorr verification by calculating the challenge and
 /// confirming the signature equation.
+///
+/// # Feature
+///
+/// Only available with the "alloc" feature enabled.
 ///
 /// ### Parameters
 ///
@@ -36,7 +43,9 @@ use dusk_plonk::prelude::*;
 ///
 /// This function will return an `Error` if the witness `u` is not a valid
 /// [`JubJubScalar`].
-pub fn single_key_verify<C: Composer>(
+///
+/// [`Signature`]: [`crate::Signature`]
+pub fn verify_signature<C: Composer>(
     composer: &mut C,
     u: Witness,
     r: WitnessPoint,
@@ -58,8 +67,12 @@ pub fn single_key_verify<C: Composer>(
     Ok(())
 }
 
-/// Verifies a [`DoubleSignature`] within a Plonk circuit without requiring
+/// Verifies a [`SignatureDouble`] within a Plonk circuit without requiring
 /// the secret key as a witness.
+///
+/// # Feature
+///
+/// Only available with the "double" and "alloc" features enabled.
 ///
 /// ### Parameters
 ///
@@ -81,8 +94,9 @@ pub fn single_key_verify<C: Composer>(
 /// This function will return an `Error` if the witness `u` is not a valid
 /// [`JubJubScalar`].
 ///
-/// [`DoubleSignature`]: [`crate::DoubleSignature`]
-pub fn double_key_verify<C: Composer>(
+/// [`SignatureDouble`]: [`crate::SignatureDouble`]
+#[cfg(feature = "double")]
+pub fn verify_signature_double<C: Composer>(
     composer: &mut C,
     u: Witness,
     r: WitnessPoint,
@@ -110,6 +124,61 @@ pub fn double_key_verify<C: Composer>(
 
     composer.assert_equal_point(r, point);
     composer.assert_equal_point(r_p, point_p);
+
+    Ok(())
+}
+
+/// Verifies a Schnorr signature with variable generator [`SignatureVarGen`]
+/// within a Plonk circuit without requiring the secret key as a witness.
+///
+/// The function performs Schnorr verification by calculating the challenge and
+/// confirming the signature equation.
+///
+/// # Feature
+///
+/// Only available with the "var_generator" and "alloc" features enabled.
+///
+/// ### Parameters
+///
+/// - `composer`: A mutable reference to the Plonk [`Composer`]`.
+/// - `u`: Witness for the random nonce used during signature generation.
+/// - `r`: Witness Point representing the nonce point `r = u*G`.
+/// - `pk`: Witness Point representing the public key `pk = sk*G`.
+/// - `gen`: Witness Point representing the variable generator `G`
+/// - `msg`: Witness for the message.
+///
+/// ### Returns
+///
+/// - `Result<(), Error>`: Returns an empty `Result` on successful gadget
+///   creation or an `Error` if the witness `u` is not a valid [`JubJubScalar`].
+///
+/// ### Errors
+///
+/// This function will return an `Error` if the witness `u` is not a valid
+/// [`JubJubScalar`].
+///
+/// [`SignatureVarGen`]: [`crate::SignatureVarGen`]
+#[cfg(feature = "var_generator")]
+pub fn verify_signature_var_gen<C: Composer>(
+    composer: &mut C,
+    u: Witness,
+    r: WitnessPoint,
+    pk: WitnessPoint,
+    gen: WitnessPoint,
+    msg: Witness,
+) -> Result<(), Error> {
+    let r_x = *r.x();
+    let r_y = *r.y();
+
+    let challenge = [r_x, r_y, msg];
+    let challenge_hash = sponge::truncated::gadget(composer, &challenge);
+
+    // TODO: check whether we need to append the generator as a constant
+    let s_a = composer.component_mul_point(u, gen);
+    let s_b = composer.component_mul_point(challenge_hash, pk);
+    let point = composer.component_add_point(s_a, s_b);
+
+    composer.assert_equal_point(r, point);
 
     Ok(())
 }