+++ tests/schnorr_double.rs
@@ -6,7 +6,8 @@
 
 use dusk_bls12_381::BlsScalar;
 use dusk_bytes::Serializable;
-use dusk_schnorr::{PublicKey, SecretKey, Signature};
+use dusk_schnorr::{PublicKeyDouble, SecretKey, SignatureDouble};
+use ff::Field;
 use rand::rngs::StdRng;
 use rand::SeedableRng;
 
@@ -15,12 +16,12 @@ fn sign_verify() {
     let mut rng = StdRng::seed_from_u64(2321u64);
 
     let sk = SecretKey::random(&mut rng);
-    let message = BlsScalar::uni_random(&mut rng);
-    let pk = PublicKey::from(&sk);
+    let message = BlsScalar::random(&mut rng);
+    let pk_double: PublicKeyDouble = sk.into();
 
-    let sig = sk.sign(&mut rng, message);
+    let sig = sk.sign_double(&mut rng, message);
 
-    assert!(sig.verify(&pk, message));
+    assert!(sig.verify(&pk_double, message));
 }
 
 #[test]
@@ -28,14 +29,15 @@ fn test_wrong_keys() {
     let mut rng = StdRng::seed_from_u64(2321u64);
 
     let sk = SecretKey::random(&mut rng);
-    let message = BlsScalar::uni_random(&mut rng);
+    let message = BlsScalar::random(&mut rng);
 
-    let sig = sk.sign(&mut rng, message);
+    let sig = sk.sign_double(&mut rng, message);
 
     // Derive random public key
-    let pk = PublicKey::from(&SecretKey::random(&mut rng));
+    let wrong_sk = SecretKey::random(&mut rng);
+    let pk_double: PublicKeyDouble = wrong_sk.into();
 
-    assert!(!sig.verify(&pk, message));
+    assert!(!sig.verify(&pk_double, message));
 }
 
 #[test]
@@ -43,8 +45,8 @@ fn to_from_bytes() {
     let mut rng = StdRng::seed_from_u64(2321u64);
 
     let sk = SecretKey::random(&mut rng);
-    let message = BlsScalar::uni_random(&mut rng);
+    let message = BlsScalar::random(&mut rng);
 
-    let sig = sk.sign(&mut rng, message);
-    assert_eq!(sig, Signature::from_bytes(&sig.to_bytes()).unwrap());
+    let sig = sk.sign_double(&mut rng, message);
+    assert_eq!(sig, SignatureDouble::from_bytes(&sig.to_bytes()).unwrap());
 }